ВАЖНО!!! Файлы conf.properties и waiter.jks не добавлять в GIT!!!
В build.gradle.kts приложения прописаны конфигурации для подписи приложения release версии, а так же созданы общедоступные параметры авторизационных данных к api, для использования внутри проекта.
Подключен ретрофит
В build.gradle.kts создан релизный конфиг для подписи пакета(в корень папки проекта нужно добавить файл waiter.jks), а так же conf.properties c конфигурационными данными для работы с API.
В Di добавлен API модуль.
А так же в модуле репозитория сделано разделение с помощью квалифаера на удаленный и замоканный репозитории. Для экрана счетов оставил пока замоканный репозиторий.
Для удаленного репозитория создан RemoteDataSourceImpl, для которого создан WaiterApi, для загрузки данных с помощью ретрофита. Подключен Interceptor, который пишет в лог в Debug режиме.
Во фрагмент счета добавлен вывод товаров счета и список групп товаров с товарами, полченных через API
Добавлен общий для экранов интерфейс IScreenView, нужно заимплементить для всех экранов (чтобы в каждом экране была возможность показать индикацию загрузки и ошибку).

+++ domain/src/main/java/com/mwaiterdev/domain/models/Item.kt
@@ -1,7 +1,7 @@
 package com.mwaiterdev.domain.models
 
 data class Item(
-    val itemId: Int,
+    val itemId: Long,
     val itemGroupId: Long,
     val name: String,
     val shorName: String,
