+++ domain/src/main/java/com/mwaiterdev/domain/models/Price.kt
@@ -1,6 +1,6 @@
 package com.mwaiterdev.domain.models
 
 data class Price(
-    val itemId: Int,
-    val price: Int
+    val itemId: Long,
+    val price: Float
 )
