+++ domain/src/main/java/com/mwaiterdev/domain/models/TableGroupItem.kt
@@ -1,7 +1,7 @@
 package com.mwaiterdev.domain.models
 
 data class TableGroupItem(
-    val tableGroupId: Int,
+    val tableGroupId: Long,
     val name: String,
     val askCustomers: Boolean,
     var isExpanded: Boolean = false
